+++ integrations/zendesk/src/channels/index.ts
@@ -1,0 +1,8 @@
+import * as bp from '../../.botpress'
+import { hitl } from './hitl'
+import { messaging } from './messaging'
+
+export default {
+  messaging,
+  hitl,
+} satisfies bp.IntegrationProps['channels']
